requirements.txt
Django>=4.2,<5.1
Pillow>=10.0
numpy>=1.24
requests>=2.31
python-decouple>=3.8
whitenoise>=6.6
